improvement: 增加L7 logo是否显示属性
