+++ packages/core/src/services/scene/ISceneService.ts
@@ -10,7 +10,7 @@ export interface ISceneService {
   addLayer(layer: ILayer): void;
   render(): void;
   getSceneContainer(): HTMLDivElement;
-  ExportMap2Png(): string;
+  exportPng(): string;
   destroy(): void;
 }
 // scene 事件
