+++ packages/maps/src/index.ts
@@ -1,4 +1,4 @@
-import GaodeMap from './amap/Wrapper';
-import Mapbox from './mapbox/Wrapper';
+import GaodeMap from './amap/';
+import Mapbox from './mapbox/';
 
 export { GaodeMap, Mapbox };
