+++ packages/maps/src/mapbox/map.ts
@@ -1,0 +1,328 @@
+/**
+ * MapboxService
+ */
+import {
+  Bounds,
+  CoordinateSystem,
+  ICoordinateSystemService,
+  IGlobalConfigService,
+  ILngLat,
+  ILogService,
+  IMapConfig,
+  IMapService,
+  IPoint,
+  IViewport,
+  MapServiceEvent,
+  MapStyle,
+  TYPES,
+} from '@antv/l7-core';
+import { DOM } from '@antv/l7-utils';
+import { inject, injectable } from 'inversify';
+import mapboxgl, { IControl, Map } from 'mapbox-gl';
+import { IMapboxInstance } from '../../typings/index';
+import Viewport from './Viewport';
+const EventMap: {
+  [key: string]: any;
+} = {
+  mapmove: 'move',
+  camerachange: 'move',
+};
+import { MapTheme } from './theme';
+
+const LNGLAT_OFFSET_ZOOM_THRESHOLD = 12;
+const MAPBOX_API_KEY =
+  'pk.eyJ1IjoibHp4dWUiLCJhIjoiYnhfTURyRSJ9.Ugm314vAKPHBzcPmY1p4KQ';
+/**
+ * AMapService
+ */
+@injectable()
+export default class MapboxService
+  implements IMapService<Map & IMapboxInstance> {
+  public map: Map & IMapboxInstance;
+
+  @inject(TYPES.MapConfig)
+  private readonly config: Partial<IMapConfig>;
+
+  @inject(TYPES.IGlobalConfigService)
+  private readonly configService: IGlobalConfigService;
+
+  @inject(TYPES.ILogService)
+  private readonly logger: ILogService;
+  @inject(TYPES.ICoordinateSystemService)
+  private readonly coordinateSystemService: ICoordinateSystemService;
+
+  @inject(TYPES.IEventEmitter)
+  private eventEmitter: any;
+  private viewport: Viewport;
+  private markerContainer: HTMLElement;
+  private cameraChangedCallback: (viewport: IViewport) => void;
+  private $mapContainer: HTMLElement | null;
+
+  // init
+  public addMarkerContainer(): void {
+    const container = this.map.getCanvasContainer();
+    this.markerContainer = DOM.create('div', 'l7-marker-container', container);
+  }
+
+  public getMarkerContainer(): HTMLElement {
+    return this.markerContainer;
+  }
+
+  //  map event
+  public on(type: string, handle: (...args: any[]) => void): void {
+    if (MapServiceEvent.indexOf(type) !== -1) {
+      this.eventEmitter.on(type, handle);
+    } else {
+      // 统一事件名称
+      this.map.on(EventMap[type] || type, handle);
+    }
+  }
+  public off(type: string, handle: (...args: any[]) => void): void {
+    this.map.off(EventMap[type] || type, handle);
+  }
+
+  public getContainer(): HTMLElement | null {
+    return this.map.getContainer();
+  }
+
+  public getSize(): [number, number] {
+    const size = this.map.transform;
+    return [size.width, size.height];
+  }
+  // get mapStatus method
+
+  public getType() {
+    return 'mapbox';
+  }
+
+  public getZoom(): number {
+    return this.map.getZoom();
+  }
+
+  public setZoom(zoom: number) {
+    return this.map.setZoom(zoom);
+  }
+
+  public getCenter(): ILngLat {
+    return this.map.getCenter();
+  }
+
+  public getPitch(): number {
+    return this.map.getPitch();
+  }
+
+  public getRotation(): number {
+    return this.map.getBearing();
+  }
+
+  public getBounds(): Bounds {
+    return this.map.getBounds().toArray() as Bounds;
+  }
+
+  public getMinZoom(): number {
+    return this.map.getMinZoom();
+  }
+
+  public getMaxZoom(): number {
+    return this.map.getMaxZoom();
+  }
+
+  public setRotation(rotation: number): void {
+    this.map.setBearing(rotation);
+  }
+
+  public zoomIn(): void {
+    this.map.zoomIn();
+  }
+
+  public zoomOut(): void {
+    this.map.zoomOut();
+  }
+
+  public panTo(p: [number, number]): void {
+    this.map.panTo(p);
+  }
+
+  public panBy(pixel: [number, number]): void {
+    this.panTo(pixel);
+  }
+
+  public fitBounds(bound: Bounds): void {
+    this.map.fitBounds(bound);
+  }
+
+  public setMaxZoom(max: number): void {
+    this.map.setMaxZoom(max);
+  }
+
+  public setMinZoom(min: number): void {
+    this.map.setMinZoom(min);
+  }
+
+  public setZoomAndCenter(zoom: number, center: [number, number]): void {
+    this.map.flyTo({
+      zoom,
+      center,
+    });
+  }
+
+  public setMapStyle(style: string): void {
+    this.map.setStyle(this.getMapStyle(style));
+  }
+  // TODO: 计算像素坐标
+  public pixelToLngLat(pixel: [number, number]): ILngLat {
+    return this.map.unproject(pixel);
+  }
+
+  public lngLatToPixel(lnglat: [number, number]): IPoint {
+    return this.map.project(lnglat);
+  }
+
+  public containerToLngLat(pixel: [number, number]): ILngLat {
+    return this.map.unproject(pixel);
+  }
+
+  public lngLatToContainer(lnglat: [number, number]): IPoint {
+    return this.map.project(lnglat);
+  }
+
+  public async init(): Promise<void> {
+    const {
+      id = 'map',
+      attributionControl = false,
+      style = 'light',
+      token = MAPBOX_API_KEY,
+      rotation = 0,
+      mapInstance,
+      ...rest
+    } = this.config;
+
+    this.viewport = new Viewport();
+
+    /**
+     * TODO: 使用 mapbox v0.53.x 版本 custom layer，需要共享 gl context
+     * @see https://github.com/mapbox/mapbox-gl-js/blob/master/debug/threejs.html#L61-L64
+     */
+
+    // 判断全局 mapboxgl 对象的加载
+    if (!mapInstance && !mapboxgl) {
+      // 用户有时传递进来的实例是继承于 mapbox 实例化的，不一定是 mapboxgl 对象。
+      this.logger.error(this.configService.getSceneWarninfo('SDK'));
+    }
+
+    if (
+      token === MAPBOX_API_KEY &&
+      style !== 'blank' &&
+      !mapboxgl.accessToken &&
+      !mapInstance // 如果用户传递了 mapInstance，应该不去干预实例的 accessToken。
+    ) {
+      this.logger.warn(this.configService.getSceneWarninfo('MapToken'));
+    }
+
+    // 判断是否设置了 accessToken
+    if (!mapInstance && !mapboxgl.accessToken) {
+      // 用户有时传递进来的实例是继承于 mapbox 实例化的，不一定是 mapboxgl 对象。
+      mapboxgl.accessToken = token;
+    }
+
+    if (mapInstance) {
+      // @ts-ignore
+      this.map = mapInstance;
+      this.$mapContainer = this.map.getContainer();
+    } else {
+      this.$mapContainer = this.creatAmapContainer(id);
+      // @ts-ignore
+      this.map = new mapboxgl.Map({
+        container: id,
+        style: this.getMapStyle(style),
+        attributionControl,
+        bearing: rotation,
+        ...rest,
+      });
+    }
+    this.map.on('load', this.handleCameraChanged);
+    this.map.on('move', this.handleCameraChanged);
+
+    // 不同于高德地图，需要手动触发首次渲染
+    this.handleCameraChanged();
+  }
+
+  public destroy() {
+    this.eventEmitter.removeAllListeners();
+    if (this.map) {
+      this.map.remove();
+      this.$mapContainer = null;
+      this.removeLogoControl();
+    }
+  }
+  public emit(name: string, ...args: any[]) {
+    this.eventEmitter.emit(name, ...args);
+  }
+  public once(name: string, ...args: any[]) {
+    this.eventEmitter.once(name, ...args);
+  }
+
+  public getMapContainer() {
+    return this.$mapContainer;
+  }
+
+  public onCameraChanged(callback: (viewport: IViewport) => void): void {
+    this.cameraChangedCallback = callback;
+  }
+
+  private handleCameraChanged = () => {
+    // @see https://github.com/mapbox/mapbox-gl-js/issues/2572
+    const { lat, lng } = this.map.getCenter().wrap();
+
+    // resync
+    this.viewport.syncWithMapCamera({
+      bearing: this.map.getBearing(),
+      center: [lng, lat],
+      viewportHeight: this.map.transform.height,
+      pitch: this.map.getPitch(),
+      viewportWidth: this.map.transform.width,
+      zoom: this.map.getZoom(),
+      // mapbox 中固定相机高度为 viewport 高度的 1.5 倍
+      cameraHeight: 0,
+    });
+
+    // set coordinate system
+    if (this.viewport.getZoom() > LNGLAT_OFFSET_ZOOM_THRESHOLD) {
+      this.coordinateSystemService.setCoordinateSystem(
+        CoordinateSystem.LNGLAT_OFFSET,
+      );
+    } else {
+      this.coordinateSystemService.setCoordinateSystem(CoordinateSystem.LNGLAT);
+    }
+
+    this.cameraChangedCallback(this.viewport);
+  };
+
+  private creatAmapContainer(id: string | HTMLDivElement) {
+    let $wrapper = id as HTMLDivElement;
+    if (typeof id === 'string') {
+      $wrapper = document.getElementById(id) as HTMLDivElement;
+    }
+    return $wrapper;
+  }
+
+  private removeLogoControl(): void {
+    // @ts-ignore
+    const controls = this.map._controls as IControl[];
+    const logoCtr = controls.find((ctr: IControl) => {
+      if (ctr.hasOwnProperty('_updateLogo')) {
+        return true;
+      }
+    });
+    if (logoCtr) {
+      this.map.removeControl(logoCtr);
+    }
+  }
+
+  private getMapStyle(name: MapStyle) {
+    if (typeof name !== 'string') {
+      return name;
+    }
+    return MapTheme[name] ? MapTheme[name] : name;
+  }
+}
